Add files via upload
Added wordify_proj.py file containing code to convert the phone number to word and vice versa. Converting a dictionary to a number and then checking for that number. Dictionary conversion is time-consuming and happens every time a conversion has to be done. 'words' from 'nltk' library is required to be installed to run the code